+++ src/PKIDisc.h
@@ -10,11 +10,17 @@
 #include "BinDisc.h"
 
 namespace mdlp {
+    enum class compute_strategy_t {
+        LOG, // Logarithmic
+        SQRT // Square root
+    };
     class PKIDisc : public BinDisc {
     public:
-        PKIDisc() = default;
+        PKIDisc(compute_strategy_t compute_strategy_ = compute_strategy_t::SQRT) : compute_strategy(compute_strategy_) {}
         ~PKIDisc() = default;
         void fit(samples_t& X_, labels_t& y) override;
+    private:
+        compute_strategy_t compute_strategy;
     };
 }
 #endif